refactor: enforce PascalCase component filenames and remove circular dependency
