+++ apps/admin/src/components/plans/PlanFormModal.tsx
@@ -17,7 +17,7 @@ import {
   planToForm,
   formToPayload,
   FieldError,
-} from "./planFormHelpers";
+} from "./PlanFormHelpers";
 import { PlanTypeLimitsSection } from "./PlanTypeLimitsSection";
 
 interface PlanFormModalProps {
